refactor: extract SessionLibraryStatusPresenter.swift from SessionLibraryPresenters.swift
Splits the `SessionLibraryStatusPresenter` @MainActor class (~36
lines) out of SessionLibraryPresenters.swift so that file holds only
the presenter enums (DisplayedTranscriptCopy* / TranscriptSave /
SessionDeletion).

Byte-identical move of lines 43-78. Both files import Combine +
Foundation.

SessionLibraryPresenters.swift: 79 → 42 lines. Tests: 667.
